+++ extensions/Splitwise/src/list-expenses.tsx
@@ -26,6 +26,7 @@ export default function Command() {
             keywords={expense.users.map((user) => user.user.first_name)}
             title={expense.description}
             accessories={[
+              { icon: expense.group_id ? Icon.TwoPeople : "", tooltip: "Group Expense" },
               {
                 tag: {
                   value: `${expense.cost} ${expense.currency_code}`,
@@ -94,7 +95,7 @@ export default function Command() {
 
                     <List.Item.Detail.Metadata.Separator />
 
-                    <List.Item.Detail.Metadata.TagList title={`Pays back`}>
+                    <List.Item.Detail.Metadata.TagList title={`Pays back `}>
                       {expense.users
                         .filter((user) => Number(user.net_balance) < 0)
                         .map((user) => (
@@ -108,6 +109,13 @@ export default function Command() {
                           />
                         ))}
                     </List.Item.Detail.Metadata.TagList>
+
+                    {expense.group_id && ( // GROUP EXPENSES
+                      <>
+                        <List.Item.Detail.Metadata.Separator />
+                        <List.Item.Detail.Metadata.Link title="Group Expense" text="View Group" target={`https://secure.splitwise.com/#/groups/${expense.group_id}`} />
+                      </>
+                    )}
 
                     {expense.repeats === true && ( // REPEATING EXPENSES
                       <>
@@ -166,12 +174,75 @@ export default function Command() {
 
 // ------------ FORM ------------
 // Comment out some lines due to updating costs not working at the moment
+
+// import { useState } from "react";
+import { useForm, FormValidation } from "@raycast/utils";
+
 function ChangeValues(handedOverValues: { expense: Expense }) {
   const { expense } = handedOverValues;
-  //   const [defaultCosts, setCosts] = useState<string>(expense.cost);
-  //   const [nameError, setNameError] = useState<string | undefined>();
+  // const [defaultCosts, setCosts] = useState<string>(expense.cost);
+  // const [nameError, setNameError] = useState<string | undefined>();
 
   const { pop } = useNavigation();
+
+  const { handleSubmit, itemProps } = useForm<Expense | any>({ //Expense
+    onSubmit: (input) => {
+      const paid_share = Number(input.cost);
+      
+      const numberShares = input.owes.length + 1;
+      // const owed_share = Number(input.cost) / numberShares;
+      
+      let share = Math.floor(Number(input.cost) * 100 / numberShares) / 100;
+      let adjustedShare = (Number(input.cost) - share) / (numberShares-1);
+      adjustedShare = Math.round(adjustedShare * 100) / 100;
+
+      const paramsJson: any  = {
+        // id: `${expense.id}` as string,
+        cost: input.cost as string,
+        description: input.description as string,
+        date: new Date(input.date).toISOString(),
+        group_id: expense.group_id as number,
+      };
+
+      paramsJson[`users__0__user_id`] = input.paid as number;
+      paramsJson[`users__0__paid_share`] = `${paid_share}`;
+      paramsJson[`users__0__owed_share`] = share;
+
+      let counter = 1;
+      input.owes.map((user:any) => {
+        paramsJson[`users__${counter}__user_id`] = user as number;
+        paramsJson[`users__${counter}__owed_share`] = adjustedShare;
+        counter++;
+      });
+
+      UpdateExpense(expense.id, paramsJson).then(() => pop());
+      // console.log(paramsJson, expense.id); pop();
+    },
+
+    initialValues: {
+      cost: expense.cost,
+      description: expense.description,
+      date: new Date(expense.date),
+      // paid: expense.users.filter((user) => Number(user.paid_share) > 0).map((user) => String(user.user.id)),
+      paid: expense.users.filter((user) => Number(user.paid_share) > 0).map((user) => String(user.user.id))[0],
+      owes: expense.users.filter((user) => Number(user.net_balance) < 0).map((user) => String(user.user.id)),
+    },
+
+    validation: {
+      description: FormValidation.Required,
+      date: FormValidation.Required,
+      paid: FormValidation.Required,
+      owes: FormValidation.Required,
+
+      cost: (input) => {
+        if (!input?.match(/^\d+(\.\d{1,2})?$/)) {
+          // check if input is integer or float with 1 or 2 decimal places
+          return "Decimal value (2 places)";
+        }
+      },
+    },
+
+  });
 
   return (
     <Form
@@ -180,37 +251,81 @@ function ChangeValues(handedOverValues: { expense: Expense }) {
         <ActionPanel>
           <Action.SubmitForm
             title="Submit changes"
-            onSubmit={(values) => {
-              UpdateExpense({
-                id: expense.id,
-                description: values.description as string,
-                // cost: values.cost as string,
-                date: values.date,
-                group_id: expense.group_id as number,
-              }).then(() => pop());
-            }}
+            onSubmit={handleSubmit}
+            // onSubmit={(values) => {
+            //   UpdateExpense({
+            //     id: expense.id,
+            //     description: values.description as string,
+            //     // cost: values.cost as string,
+            //     date: values.date,
+            //     group_id: expense.group_id as number,
+            //   }).then(() => pop());
+            // }}
           />
         </ActionPanel>
       }
     >
-      <Form.TextField title="Description" id="description" defaultValue={expense.description} />
-      {/* <Form.TextField
-        title={`Costs in ${expense.currency_code}`}
-        id="cost"
-        onChange={setCosts}
-        value={defaultCosts}
-        error={nameError}
-        onBlur={(input) => {
-          if (!input.target.value?.match(/^\d+(\.\d{1,2})?$/)) {
-            // check if input is integer or float with 1 or 2 decimal places
-            setNameError("The field should't be empty!");
-          } else {
-            setCosts;
-            setNameError(undefined);
-          }
-        }}
-      /> */}
-      <Form.DatePicker title="Date of Expense" id="date" defaultValue={new Date(expense.date)} />
+      <Form.TextField
+        title="Description"
+        {...itemProps.description} //id="description" // defaultValue={expense.description}
+      />
+      <Form.TextField
+        title={`Cost in ${expense.currency_code}`}
+        {...itemProps.cost}
+        // id="cost"
+        // onChange={setCosts}
+        // value={defaultCosts}
+        // error={nameError}
+        // onBlur={(input) => {
+        //   if (!input.target.value?.match(/^\d+(\.\d{1,2})?$/)) {
+        //     // check if input is integer or float with 1 or 2 decimal places
+        //     setNameError("The field should't be empty!");
+        //   } else {
+        //     setCosts;
+        //     setNameError(undefined);
+        //   }
+        // }}
+        // defaultValue={expense.cost}
+      />
+      <Form.DatePicker
+        title="Date of Expense"
+        {...itemProps.date} //id="date" defaultValue={new Date(expense.date)}
+      />
+
+      <Form.Separator />
+
+      <Form.Dropdown
+        title="Who paid?"
+        {...itemProps.paid}
+        // info="If multiple people selected, the paid share will be split equally"
+      >
+        {expense.users.map((user) => (
+          <Form.Dropdown.Item
+            key={user.user.id}
+            value={String(user.user.id)}
+            title={[user.user.first_name, user.user.last_name].join(" ")}
+            icon={{ source: user.user.picture.medium, mask: Image.Mask.Circle}}
+          />
+        ))}
+      </Form.Dropdown>
+
+      <Form.TagPicker
+        title="Who owes?"
+        {...itemProps.owes}
+        info="Expense will be split equally among the involved people"
+        // id="owes"
+        // defaultValue={expense.users.filter((user) => Number(user.net_balance) < 0).map((user) => String(user.user.id))}
+      >
+        {expense.users.map((user) => (
+          <Form.TagPicker.Item
+            key={user.user.id}
+            value={String(user.user.id)}
+            title={user.user.first_name}
+            icon={{ source: user.user.picture.medium, mask: Image.Mask.Circle}}
+          />
+        ))}
+      </Form.TagPicker>
+
     </Form>
   );
 }